revert double  to uint64_t, bigger would entail extreme calculation periods. further async experiments refractoring. added gets in fraction

+++ PrimeFactorizer/PrimeCheck.h
@@ -9,7 +9,7 @@ class PrimeCheck
 public:
 	PrimeCheck() noexcept; 
 	PrimeCheck(shared_ptr <PrimeLibrary> l);
-	bool is_prime(double input);
+	bool is_prime(uint64_t input);
 
 private:
 	shared_ptr <PrimeLibrary> Lib;
